V2: Full functional FE with Recharts + transcripts integrated
This commit contains the integration of Recharts to process the final data returned from api/races/driver_data/{driver_name} so that it can be visualized. It also contains rendering of the transcripts when data points are clicked on the chart, as well as a fully functional user input flow, load state, and then loading of the sim data on the chart.

I also made sure to add backup data in case the driver_data ever errors out, so that the user can still visualize what we've worked on.

The web app is also mobile friendly.